locks: use cmpxchg to assign i_flctx pointer
During the v3.20/v4.0 cycle, I had originally had the code manage the
inode->i_flctx pointer using a compare-and-swap operation instead of
the i_lock.

Sasha Levin though hit a problem while testing trinity that made me
believe that that wasn't safe. I now think though that I completely
misread the problem, even though it seemed like it went away when
we started using the i_lock to protect this pointer.

The issue was likely the same race that Kirill Shutemov hit while
testing the pre-rc1 v4.0 kernel and that Linus spotted. Due to the
way that the spinlock was dropped in the middle of flock_lock_file,
you could end up with multiple flock locks for the same struct file
on the inode.

Reinstate the use of a CAS operation to assign this pointer since it's
likely to be more efficient and gets the i_lock completely out of the
file locking business.

Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>