thermal: user_space: add trip point number to uevent data
A user-space thermal governor does not currently know which trip point
caused a uevent to trigger. Let's add this information to the payload.

This is just an RFC, because this patch (on its own) is not very useful.
The only current caller of the .throttle callback is in
thermal_zone_device_update(), which loops over *all* trip points. This
seems very counter-intuitive and unhelpful, but maybe I'm just clueless.

Signed-off-by: Brian Norris <computersforpeace@gmail.com>
Signed-off-by: Zhang Rui <rui.zhang@intel.com>