nfsd4: disallow ALLOCATE with special stateids
I haven't tested, but believe vfs_fallocate will hit a NULL dereference
if the client tries an ALLOCATE or DEALLOCATE with a special stateid.
Fix that.  (We also depend on the open to have broken any conflicting
leases or delegations for us.)

(If it turns out we need to allow special stateid's then we could do a
temporary open here in the special-stateid case, as we do for read and
write.  For now I'm assuming it's not necessary.)

Cc: stable@vger.kernel.org
Signed-off-by: J. Bruce Fields <bfields@redhat.com>