target: simplify the target template registration API
On Tue, Apr 07, 2015 at 11:12:50PM -0700, Nicholas A. Bellinger wrote:
> So I was going to just merge this because having a single register
> template function for fabric drivers does make alot of sense, but it
> currently conflicts with the T10-PI fabric bits in for-next.  Sigh..
>
> Looking further, I really don't like how the target_fabric_copy_attrs()
> stuff works in target_register_template().  Just punting to direct use
> struct configfs_attributes pointers in target_core_fabric_ops is a hack.
>
> I'd much rather see target_fabric_setup_cits() do this proper setup,
> instead of unnecessarily duplicating the target_core_fabric_configfs.c
> list of defined config_item_types in two new locations.

Does this version look better?  It adds a new TF_CIT_SETUP_DRV macro,
so that the core configfs code can declare attributes as either core
only or for drivers.  That way we also avoid the case where the core
code defines an attribute and a fabric driver accidentally overrides it.

It's also been rebased to latest target/for-next.

Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>