EVM: labeling psuedo filesystems exception
To prevent offline stripping of existing file xattrs and relabeling of
them at runtime, EVM allows only newly created files to be labeled.  As
psuedo filesystems are not persistent, stripping of xattrs is not a
concern.

Some LSMs defer file labeling on psuedo filesystems.  This patch
permits the labeling of existing files on psuedo files systems.

Signed-off-by: Mimi Zohar <zohar@linux.vnet.ibm.com>