clk: prevent orphan clocks from being used
Orphan clocks or children of orphan clocks don't have rate
information at all and can produce strange results if they're
allowed to be used and the parent becomes available later on. Use
the newly introduced orphan status bit to defer clk_get() callers
if the clock is an orphan.

Based on a patch from Heiko Stuebner.

Reviewed-by: Heiko Stuebner <heiko@sntech.de>
Tested-by: Heiko Stuebner <heiko@sntech.de>
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>