soc/tegra: Add EMC API used by clock driver
The EMC clock driver needs to program the external memory controller at
a very specific time in the rate transition sequence in order to ensure
memory accesses to continue to work.

This commit provides the corresponding function prototypes that will be
exposed by the EMC driver. Dummy implementations are provided for build
configurations where the EMC driver is disabled. This is also used as a
means to decouple both drivers so that they can be merged independently
and don't cause build breakage.

Signed-off-by: Thierry Reding <treding@nvidia.com>