include/linux/sched.h: change all uses of JOBCTL_* from int to long
c56fb6564dcd ("Fix a misaligned load inside ptrace_attach()") makes jobctl
an "unsigned long".  It makes sense to have the masks applied to it match
that type.  This is currently just a cosmetic change, but it will prevent
the mask from being unexpectedly truncated if we ever end up with masks
with more bits.

One instance of "signr" is an int, but I left this alone because the mask
ensures that it will never overflow.

Signed-off-by: Palmer Dabbelt <palmer@dabbelt.com>
Reviewed-by: Chris Metcalf <cmetcalf@ezchip.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Oleg Nesterov <oleg@redhat.com>
Cc: Richard Weinberger <richard@nod.at>
Cc: "Paul E. McKenney" <paulmck@linux.vnet.ibm.com>
Cc: Pranith Kumar <bobby.prani@gmail.com>
Cc: Vladimir Davydov <vdavydov@parallels.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>