x86/build: Remove -Wno-sign-compare
Louis Langholtz noticed that we have the -Wno-sign-compare compiler option
in arch/x86/Makefile, which turns off certain, useful looking warnings.

This quirk is documented as:

   # Workaround for a gcc prelease that unfortunately was shipped in a suse release

The workaround is more than 10 years old, it was added in a large x86_64
merge in v2.5.70:

   47e4079c7da4 ("[PATCH] x86-64 merge")

(as per tglx's historic Linux Git tree.)

Probably described by line in the changelog:

   - Disable sign compare warnings for gcc 3.3-hammer

That's ancient and we generate quite a few warnings with old
GCC versions in any case. Worst case we such an ancient GCC
will generate a few more warnings.

So restore the warning and see what happens.

Reported-by: Louis Langholtz <lou_langholtz@me.com>
Cc: Andy Lutomirski <luto@amacapital.net>
Cc: Borislav Petkov <bp@alien8.de>
Cc: Brian Gerst <brgerst@gmail.com>
Cc: Denys Vlasenko <dvlasenk@redhat.com>
Cc: H. Peter Anvin <hpa@zytor.com>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Ingo Molnar <mingo@kernel.org>