iommu/vt-d: Add functions to load and save old root entry
Add functions to load root entry table from old kernel, and
to save updated root entry table.  Add two member in struct
intel_iommu, to store the RTA in old kernel, and the mapped
virt address of it.

We use the old RTA in dump kernel, and when the
iommu->root_entry is used as a cache in kdump kernel, its
phys address will not be save to RTA register, but when its
data is changed, we will save the new data to old root entry
table.

Li, Zhen-hua:
    The functions and logics.

Takao Indoh:
    Add __iommu_flush_cache.

Signed-off-by: Li, Zhen-Hua <zhen-hual@hp.com>
Signed-off-by: Takao Indoh <indou.takao@jp.fujitsu.com>
Signed-off-by: Joerg Roedel <jroedel@suse.de>