f2fs: truncate blocks if inode remains in the cache
As Hu reported, F2FS has a space leak problem, when conducting:

1) format a 4GB f2fs partition
2) dd a 3G file,
3) unlink it.

And, it turns out the problem is due to the bypassing dropping the inode by
f2fs_drop_inode.

In f2fs_drop_inode, if the inode is unlinked, we don't need to skip eviction.
Instead, we can truncate all the blocks in prior to eviction.

Reported-by: hujianyang <hujianyang@huawei.com>
Signed-off-by: Jaegeuk Kim <jaegeuk@kernel.org>