mnt: Modify fs_fully_visible to deal with locked nosuid and noexec
Ignore an existing mount if it is locked nosuid and locked noexec
attributes are less permissive than the new desired attributes of the
new mount.

On success ensure the new mount locks all of the same nosuid and
noexec attributes as the old mount.

This change interacts with bugs in starting unprivileged containers in
lxc and libvirt-lxc.  Unfortunately for those programs this is a
security and maintenance issue.

In the next couple of decades I fully expect someone will either come
up with a good reason to have an executable in proc or sysfs or to
implement a buggy version of chmod again that allows someone to
convert an existing proc or sysfs file into a suid root exectuable.
That is in the fullness of time I expect not enforcing nosuid and
noexec to become exploitable.

Therefore start requiring nosuid or noexec if the original mount of
proc or sysfs also had those flags set.  This keeps an unprivileged
user from violating the global policy that was set when proc or sysfs
was initially mounted.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>